CR comments (newest/current version distinction + renames).
