+++ core/src/main/java/com/radixdlt/p2p/P2PModule.java
@@ -85,13 +85,19 @@ import com.radixdlt.p2p.hostip.HostIpModule;
 import com.radixdlt.p2p.transport.PeerOutboundBootstrap;
 import com.radixdlt.p2p.transport.PeerOutboundBootstrapImpl;
 import com.radixdlt.p2p.transport.PeerServerBootstrap;
-import com.radixdlt.p2p.transport.ProtocolUpdateId;
+import com.radixdlt.protocol.Current;
+import com.radixdlt.protocol.Newest;
+import com.radixdlt.protocol.ProtocolVersion;
 import com.radixdlt.serialization.Serialization;
 import com.radixdlt.store.BerkeleyDbDefaults;
 import com.radixdlt.store.NodeStorageLocation;
 import com.radixdlt.utils.properties.RuntimeProperties;
 
 public final class P2PModule extends AbstractModule {
+
+  /** The only currently defined protocol version. */
+  private static final ProtocolVersion ONLY_PROTOCOL_VERSION = new ProtocolVersion("babylon-v1");
+
   private final RuntimeProperties properties;
 
   public P2PModule(RuntimeProperties properties) {
@@ -107,7 +113,8 @@ public final class P2PModule extends AbstractModule {
     eventBinder.addBinding().toInstance(PeerOutboundConnectionTimeout.class);
 
     // TODO(when introducing actual protocol updates): design how to manage this
-    bind(ProtocolUpdateId.class).toInstance(new ProtocolUpdateId("babylon-v1"));
+    bind(ProtocolVersion.class).annotatedWith(Current.class).toInstance(ONLY_PROTOCOL_VERSION);
+    bind(ProtocolVersion.class).annotatedWith(Newest.class).toInstance(ONLY_PROTOCOL_VERSION);
 
     bind(AddressBook.class).in(Scopes.SINGLETON);
     bind(PeersView.class).to(PeerManagerPeersView.class).in(Scopes.SINGLETON);
