+++ core/src/test/java/com/radixdlt/api/SystemApiTestBase.java
@@ -91,7 +91,8 @@ import com.radixdlt.p2p.P2PConfig;
 import com.radixdlt.p2p.RadixNodeUri;
 import com.radixdlt.p2p.TestP2PModule;
 import com.radixdlt.p2p.addressbook.AddressBook;
-import com.radixdlt.p2p.transport.ProtocolUpdateId;
+import com.radixdlt.protocol.Current;
+import com.radixdlt.protocol.ProtocolVersion;
 import com.radixdlt.rev2.Decimal;
 import com.radixdlt.rev2.modules.REv2StateManagerModule;
 import com.radixdlt.sync.SyncRelayConfig;
@@ -142,7 +143,9 @@ public abstract class SystemApiTestBase {
             new AbstractModule() {
               @Override
               protected void configure() {
-                bind(ProtocolUpdateId.class).toInstance(new ProtocolUpdateId("test"));
+                bind(ProtocolVersion.class)
+                    .annotatedWith(Current.class)
+                    .toInstance(new ProtocolVersion("test"));
                 bind(Network.class).toInstance(Network.INTEGRATIONTESTNET);
                 bind(P2PConfig.class).toInstance(p2pConfig);
                 bind(AddressBook.class).in(Scopes.SINGLETON);
